# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdapterFlow",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/marketplaces/overview": {
      "get": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Get Marketplaces Overview",
        "description": "Retorna o panorama real das integrações de marketplaces e contas conectadas.",
        "operationId": "get_marketplaces_overview_api_v1_marketplaces_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplacesOverviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/v1/marketplaces/accounts": {
      "get": {
        "tags": [
          "marketplaces"
        ],
        "summary": "List Marketplace Accounts",
        "description": "Lista as contas conectadas de marketplaces.",
        "operationId": "list_marketplace_accounts_api_v1_marketplaces_accounts_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "marketplace",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Marketplace"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketplaceAccountResponse"
                  },
                  "title": "Response List Marketplace Accounts Api V1 Marketplaces Accounts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplaces/accounts/{id}": {
      "delete": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Disconnect Account",
        "description": "Desconecta uma conta de marketplace.",
        "operationId": "disconnect_account_api_v1_marketplaces_accounts__id__delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplaces/mercadolivre/auth-url": {
      "get": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Get Mercadolivre Auth Url",
        "operationId": "get_mercadolivre_auth_url_api_v1_marketplaces_mercadolivre_auth_url_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/v1/marketplaces/mercadolivre/oauth/callback": {
      "post": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Mercadolivre Oauth Callback",
        "description": "Processa o código recebido do Mercado Livre e conclui a autenticação da conta.",
        "operationId": "mercadolivre_oauth_callback_api_v1_marketplaces_mercadolivre_oauth_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceAccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/v1/marketplaces/mercadolivre/categories/predict": {
      "get": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Predict Category",
        "description": "Sugere categorias no Mercado Livre a partir do título do produto.",
        "operationId": "predict_category_api_v1_marketplaces_mercadolivre_categories_predict_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 60,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryPredictionItem"
                  },
                  "title": "Response Predict Category Api V1 Marketplaces Mercadolivre Categories Predict Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplaces/mercadolivre/publish": {
      "post": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Publish To Mercadolivre",
        "description": "Publica um produto no Mercado Livre com validação prévia e preço da Fase 2.",
        "operationId": "publish_to_mercadolivre_api_v1_marketplaces_mercadolivre_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/v1/marketplaces/listings": {
      "get": {
        "tags": [
          "marketplaces"
        ],
        "summary": "List Marketplace Listings",
        "description": "Lista anúncios e publicações nos marketplaces.",
        "operationId": "list_marketplace_listings_api_v1_marketplaces_listings_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceListingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplaces/mercadolivre/categories/{category_id}/attributes": {
      "get": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Category Attributes",
        "operationId": "category_attributes_api_v1_marketplaces_mercadolivre_categories__category_id__attributes_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category Id"
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplaces/listings/{listing_id}/reconcile": {
      "post": {
        "tags": [
          "marketplaces"
        ],
        "summary": "Reconcile Listing",
        "operationId": "reconcile_listing_api_v1_marketplaces_listings__listing_id__reconcile_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Listing Id"
            }
          },
          {
            "name": "external_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^MLB[0-9]+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "External Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryPredictionItem": {
        "properties": {
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "domain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "domain_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Name"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "category_name"
        ],
        "title": "CategoryPredictionItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketplaceAccountResponse": {
        "properties": {
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "connection_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Error"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "marketplace": {
            "type": "string",
            "title": "Marketplace"
          },
          "account_name": {
            "type": "string",
            "title": "Account Name"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "token_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Token Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "marketplace",
          "account_name",
          "seller_id",
          "site_id",
          "is_active",
          "token_expires_at",
          "created_at"
        ],
        "title": "MarketplaceAccountResponse"
      },
      "MarketplaceChannelStatus": {
        "properties": {
          "marketplace": {
            "type": "string",
            "title": "Marketplace"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_configured": {
            "type": "boolean",
            "title": "Is Configured"
          },
          "is_connected": {
            "type": "boolean",
            "title": "Is Connected"
          },
          "accounts_count": {
            "type": "integer",
            "title": "Accounts Count"
          },
          "auth_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Url"
          }
        },
        "type": "object",
        "required": [
          "marketplace",
          "name",
          "is_configured",
          "is_connected",
          "accounts_count"
        ],
        "title": "MarketplaceChannelStatus"
      },
      "MarketplaceListingListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MarketplaceListingResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "MarketplaceListingListResponse"
      },
      "MarketplaceListingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Name"
          },
          "marketplace": {
            "type": "string",
            "title": "Marketplace"
          },
          "external_listing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Listing Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "available_quantity": {
            "type": "integer",
            "title": "Available Quantity"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "listing_type_id": {
            "type": "string",
            "title": "Listing Type Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "permalink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permalink"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "account_id",
          "marketplace",
          "title",
          "price",
          "available_quantity",
          "category_id",
          "listing_type_id",
          "status",
          "created_at"
        ],
        "title": "MarketplaceListingResponse"
      },
      "MarketplacesOverviewResponse": {
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/MarketplaceChannelStatus"
            },
            "type": "array",
            "title": "Channels"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/MarketplaceAccountResponse"
            },
            "type": "array",
            "title": "Accounts"
          }
        },
        "type": "object",
        "required": [
          "channels",
          "accounts"
        ],
        "title": "MarketplacesOverviewResponse"
      },
      "OAuthCallbackRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "state": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "code",
          "state"
        ],
        "title": "OAuthCallbackRequest"
      },
...
      "PublishProductRequest": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "pricing_profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pricing Profile Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "title": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Title"
          },
          "category_id": {
            "type": "string",
            "pattern": "^MLB[0-9]+$",
            "title": "Category Id"
          },
          "listing_type_id": {
            "type": "string",
            "enum": [
              "gold_special",
              "gold_pro"
            ],
            "title": "Listing Type Id"
          },
          "available_quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Available Quantity"
          },
          "condition": {
            "type": "string",
            "enum": [
              "new",
              "used",
              "not_specified"
            ],
            "title": "Condition"
          },
          "attributes": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "account_id",
          "pricing_profile_id",
          "request_id",
          "title",
          "category_id",
          "listing_type_id",
          "available_quantity",
          "condition"
        ],
        "title": "PublishProductRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}